Changed LineWidth from float64 to float32 to match others

+++ gl_webgl.go
@@ -905,8 +905,8 @@ func (c *Context) IsEnabled(capability int) bool {
 }
 
 // Sets the width of lines in WebGL.
-func (c *Context) LineWidth(width float64) {
-	c.Call("lineWidth", width)
+func (c *Context) LineWidth(width float32) {
+	c.Call("lineWidth", float64(width))
 }
 
 // Links an attached vertex shader and an attached fragment shader
